+++ fibjs/include/date.h
@@ -87,18 +87,18 @@ public:
     template <typename ts>
     void set_timestamp(const ts& v)
     {
-        d = (double)v.tv_sec * 1000ll + (double)v.tv_nsec / 1000000.0;
+        d = (double)v.tv_sec * 1000 + (double)v.tv_nsec / 1000000.0;
     }
 
     template <typename ts>
     void get_timestamp(ts& v)
     {
-        v.tv_sec = floor(d / 1000);
-        v.tv_nsec = (d - v.tv_sec * 1000ll) * 1000000ll;
+        v.tv_sec = (int64_t)floor(d / 1000);
+        v.tv_nsec = (uint32_t)((d - v.tv_sec * 1000) * 1000000);
 
-        int64_t sec1 = floor((double)v.tv_nsec / 1000000000ll);
+        int32_t sec1 = (int32_t)floor((double)v.tv_nsec / 1000000000);
         v.tv_sec += sec1;
-        v.tv_nsec -= sec1 * 1000000000ll;
+        v.tv_nsec -= (uint32_t)(sec1 * 1000000000);
     }
 
     v8::Local<v8::Value> value(v8::Isolate* isolate) const
