+++ fibjs/src/db/sql/db_api.h
@@ -161,7 +161,7 @@ result_t db_begin(T* pThis, exlib::string point, AsyncEvent* ac, bool mssql = fa
         return pThis->execute("BEGIN TRANSACTION", 17, retVal);
     else {
         exlib::string str((mssql ? "SAVE TRANSACTION " : "SAVEPOINT ") + point);
-        return pThis->execute(str.c_str(), str.length(), retVal);
+        return pThis->execute(str.c_str(), (int32_t)str.length(), retVal);
     }
 }
 
@@ -179,10 +179,10 @@ result_t db_commit(T* pThis, exlib::string point, AsyncEvent* ac, bool mssql = f
     if (point.empty())
         return pThis->execute("COMMIT", 6, retVal);
     else {
-        if(mssql)
+        if (mssql)
             return 0;
         exlib::string str((mssql ? "COMMIT TRANSACTION " : "RELEASE SAVEPOINT ") + point);
-        return pThis->execute(str.c_str(), str.length(), retVal);
+        return pThis->execute(str.c_str(), (int32_t)str.length(), retVal);
     }
 }
 
@@ -201,7 +201,7 @@ result_t db_rollback(T* pThis, exlib::string point, AsyncEvent* ac, bool mssql =
         return pThis->execute("ROLLBACK", 8, retVal);
     else {
         exlib::string str((mssql ? "ROLLBACK TRANSACTION " : "ROLLBACK TO ") + point);
-        return pThis->execute(str.c_str(), str.length(), retVal);
+        return pThis->execute(str.c_str(), (int32_t)str.length(), retVal);
     }
 }
